fix(discord-bot): extract parseRerollId to avoid Bun mock interference
Bun 1.3.10 mock.module breaks module resolution for all exports in a
file that imports discord.js, even for functions with no discord.js
dependency. Extract parseRerollId into its own file (parseRerollId.ts)
and re-export from rollButton.ts for backwards compatibility.
