+++ apps/discord-bot/src/utils/rollButton.ts
@@ -1,5 +1,7 @@
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js'
 
+export { parseRerollId } from './parseRerollId.js'
+
 export function createRollButton(
   command: string,
   params: string,
@@ -13,14 +15,4 @@ export function createRollButton(
     .setDisabled(disabled)
 
   return new ActionRowBuilder<ButtonBuilder>().addComponents(button)
-}
-
-export function parseRerollId(customId: string): { command: string; params: string } | null {
-  if (!customId.startsWith('reroll:')) return null
-  const withoutPrefix = customId.slice('reroll:'.length)
-  const colonIndex = withoutPrefix.indexOf(':')
-  if (colonIndex === -1) return null
-  const command = withoutPrefix.slice(0, colonIndex)
-  const params = withoutPrefix.slice(colonIndex + 1)
-  return { command, params }
 }
